Fixes #7 - trap keyboard focus when inside drawer
This took a few hours to get right, but basically ensures that you can
only navigate between MDL navigation items inside the drawer when its
open rather than being able to tab back outside of this container. I’ve
throttled the implementation to only work on keydown when the drawer is
visible so this check isn’t happening on every keystroke.